support and enable comment previews
Support comment previews by using jQuery, the AJAX backend
added in d51715ac and some central CSS classes and styling.
Put it in a new comments.js file and add a function to en-
able it. Enable it for reviews and candidates.

Also add HTML5 "placeholder" attributes for comment input
<textarea> elements.