adjust paths to file movings
The paths passed to require() and require_once() the moved
files have been updated. Also, such calls inside the moved
files needed a fix. And last but not least, the references
to the internal URLs have been corrected by this commit.

NOTE:
In the moved files, some require_once() paths needed to be
prefixed with the directory path containing them, because
when required, the working directory is different. And for
PHP < 5.2 support, __DIR__ could not be used.